requirements.txt
pygame==1.9.6
requests
Pillow
paho-mqtt
